msm: mdss: fix QSEED3 bypass mode and lut cfg issues
This change enables the QSEED3 bypass mode in driver, until now if
scaling was not required driver was falling back to default configuration.
This change checks if scaling is not required on VIG pipes, driver only
programs the pixel extension and configures the bypass mode in QSEED3.
This change also corrects the LUT configurations for separable and
circular filters.

Change-Id: I17c8c3e422fd62feaada978ac2be48f891417db0
Signed-off-by: Abhijit Kulkarni <kabhijit@codeaurora.org>
CRs-Fixed: 1036019, 1039433
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>