input: misc: fpc1145: Add AWAKE ioctl to extend wakeup timeout.
The fingerprint hal blocks/loops for as long a finger is touching the
sensor before going back to polling. If an object stays on the sensor
too long, the wakelock times out before the HAL has decided to wake up
the system, causing it to freeze up before reaching the polling state.
The wake-enabled IRQ is only able to wake the process in that polling
state.
With this change the HAL is able to signal that it needs a little longer
before freezing up the CPU. This explicit approach has been chosen over
an infinite (pm_keep_awake) wakelock, which needs a synchronisation
point to release the wakelock. If this is not called, the system never
goes back to sleep.

Change-Id: Ib53fd96d70b030cd7c2d02da0bb8db180589daea
Signed-off-by: MarijnS95 <marijns95@gmail.com>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>