input: misc: fpc1145: Replace global drvdata pointer with private data.
Using private_data fields instead of a global variable is necessary if
multiple devices are ever supposed to be using the same driver. Even
though this is not planned, reducing global state in favor of local
state scoped to a specific device is preferred. Of course, this
per-device datastructure should not be the drvdata for the entire
platform_driver.

Change-Id: I082fef63eefa8ec1631c98e6d7fcc47b29a235eb
Signed-off-by: MarijnS95 <marijns95@gmail.com>