leds: leds-qpnp-flash: Remove unneeded features
Instead of converting the driver to the IIO API, remove the
features requiring this conversion, since they are not used
in any of the Sony Open Devices supported devices.

Anyway, if you are a developer and you're looking forward
to convert the removed features instead, I'm still giving
you the solution. Simply look at this commit:
d1cb8315040727e03031137a86874d2825165a09

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>