clk: qcom: gcc: Add global clock controller driver for SDM660
Add the clocks supported in global clock controller which clock the
peripherals like BLSPs, SDCC, USB, MDSS etc. Register all the clocks
to the clock framework for the clients to be able to request for them.

Please note that this version of the driver does not contain the debug
clock controller. It will be added later in a separate driver.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>