clk: qcom: mdss: Update DSI 14nm PLL clock names
This change updates the parent name strings for various
DSI 14nm PLL clocks. bi_tcxo is not present on SDM630 SoC
and we should use xo_board fixed-clock.
On 4.14 kernel from where this driver was taken, we will
fix this by adding a bi_tcxo fixed-clock in dts or in gcc
driver, so that the driver remains compatible with other
SoCs in which it is used. We didn't do this on 4.4, because
this driver is only used for SDM630.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>