power: qpnp-fg: Reduce debug spam in the kernel log
Add QPNP_FG_EXTENSION_DEBUG Kconfig option for to
display extended logs and debug driver information.

Also change QPNP_FG_EXTENSION to bool. The functionality
that turned on by this option is fully contained in the
driver on which it depends and always is a part of it, so
it doesn't make sense to make it tristate.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>