i2c-msm-v2.h: Correct the BRST_BW vote value for i2c
Modify the BRST_BW vote value for i2c by correcting the bus width
to 4 instead of 8 and also decrease the frequency by 1 so that
we will be voting for less than 19.2MHz. clk_freq_in is always
19.2MHz in case of i2c as the i2c clock requirement is less than
1 MHz (100KHz, 400KHz, 1MHz). So decreasing it by 1 will ensure we
are voting for less than 19.2 MHz and this will help in the case
where i2c driver goes to suspend as a part of system_suspend_noirq
call.

The issue is when i2c driver is suspending as a part of suspend
noirq call, the band width which i2c is unvoting is greater than
19.2MHz and this will require RPM acknowledgment which is not
coming in case of single core target.

Change-Id: I181957e0698590153be6200a1efd5f47c31f7b19
Signed-off-by: Prudhvi Yarlagadda <pyarlaga@codeaurora.org>
Signed-off-by: Haribabu Gattem <haribabu@codeaurora.org>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>